Create monitor_user_logins_for_intruders.sh
We can write a shell script that can scan through the logﬁ les and gather the required 
information from them. Here, we are dealing with SSH login failures. The user authentication 
session log is written to the log ﬁ le /var/log/auth.log. The script should scan the log ﬁ le 
to detect the failure login attempts and perform different checks on the log to infer the data. 
We can use the host command  to ﬁ nd out the host mapping from the IP address.